--- backend/src/server.js
@@ -1,13 +1,0 @@
-// init
-const express = require('express');
-const app = express();
-
-const PORT = 5000;
-
-app.get('/', (req, res)=>{
-    res.send('wsup')
-})
-
-app.listen(PORT, ()=>{
-    console.log('Server running on port ${port}');
-})
