+++ frontend/views/index.ejs
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title> Entry </title>
+    </head>
+    <body>
+        <h1> Entry </h1>
+    </body>
+</html>
