+++ frontend/views/login.ejs
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <script src="https://accounts.google.com/gsi/client" async defer></script>
+    <div id="g_id_onload"
+        data-client_id="<%= clientId %>"
+        data-callback="handleCredentialResponse">
+    </div>
+    <div class="g_id_signin" data-type="standard"></div>
+
+    <title>Login</title>
+</head>
+    <body>
+        <h1>Login</h1>
+
+        <script>
+            function handleCredentialResponse(response) {
+                console.log("Signed in")
+                const responsePayload = decodeJwtResponse(response.credential);
+
+                console.log("ID: " + responsePayload.sub);
+                console.log('Full Name: ' + responsePayload.name);
+                console.log('Given Name: ' + responsePayload.given_name);
+                console.log('Family Name: ' + responsePayload.family_name);
+                console.log("Image URL: " + responsePayload.picture);
+                console.log("Email: " + responsePayload.email);
+
+                // direct back to entry
+                window.location.href = '/'
+
+                // @TODO: update database if new user
+            }
+            function decodeJwtResponse(token) {
+                // Split the JWT into its three parts: header, payload, and signature
+                const base64Url = token.split('.')[1];
+                // Decode the Base64 URL-safe string to Base64
+                const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
+                // Decode the Base64 encoded payload to JSON
+                const jsonPayload = atob(base64);
+                // Parse the JSON string to get a JavaScript object
+                return JSON.parse(jsonPayload);
+            }
+        </script>
+    </body>
+</html>
